Split out dechunkifying from deserialize_get_changes_since_response.
